refactor: move preferences to separate module
